fix(lib/{helpers,api}): fix runtime error caused by usage of server side  on client side
